+++ src/api/api.js
@@ -98,7 +98,9 @@ export const filtersReducer = (acc, filter = {}) => ({
     ...'registeredWithFilter' in filter && { registeredWithFilter: filter.registeredWithFilter },
     ...'osFilter' in filter && { osFilter: filter.osFilter },
     ...'rhcdFilter' in filter && { rhcdFilter: filter.rhcdFilter },
-    ...'updateMethodFilter' in filter && { updateMethodFilter: filter.updateMethodFilter }
+    ...'lastSeenFilter' in filter && { lastSeenFilter: filter.lastSeenFilter },
+    ...'updateMethodFilter' in filter && { updateMethodFilter: filter.updateMethodFilter },
+    ...'groupHostFilter' in filter && { groupHostFilter: filter.groupHostFilter }
 });
 
 export async function getEntities(items, {
@@ -109,9 +111,10 @@ export async function getEntities(items, {
     page,
     orderBy,
     orderDirection,
-    fields = { system_profile: ['operating_system'] },
+    fields = { system_profile: ['operating_system', /* needed by inventory groups */ 'system_update_method'] },
     ...options
 }, showTags) {
+
     if (hasItems && items?.length > 0) {
         let data = await hosts.apiHostGetHostById(
             items,
@@ -180,8 +183,8 @@ export async function getEntities(items, {
             orderDirection,
             filters.staleFilter,
             [
-                ...constructTags(filters.tagFilters),
-                ...options.tags || []
+                ...constructTags(filters?.tagFilters),
+                ...options?.globalFilter?.tags || []
             ],
             filters?.registeredWithFilter,
             undefined,
@@ -189,9 +192,12 @@ export async function getEntities(items, {
             {
                 cancelToken: controller && controller.token,
                 query: {
+                    ...(options?.globalFilter?.filter && generateFilter(options.globalFilter.filter)),
                     ...(options.filter && Object.keys(options.filter).length && generateFilter(options.filter)),
                     ...(calculateSystemProfile(filters)),
-                    ...(fields && Object.keys(fields).length && generateFilter(fields, 'fields'))
+                    ...(fields && Object.keys(fields).length && generateFilter(fields, 'fields')),
+                    ...filters?.lastSeenFilter?.updatedStart && { updated_start: filters.lastSeenFilter.updatedStart },
+                    ...filters?.lastSeenFilter?.updatedEnd && { updated_end: filters.lastSeenFilter.updatedEnd }
                 }
             }
         )