+++ src/components/InventoryGroups/Modals/CreateGroupModal.js
@@ -13,7 +13,9 @@ import awesomeDebouncePromise from 'awesome-debounce-promise';
 const CreateGroupModal = ({
     isModalOpen,
     setIsModalOpen,
-    reloadData
+    reloadData,
+    modalBefore = false,
+    setterOfModalBefore
 }) => {
     const dispatch = useDispatch();
 
@@ -46,11 +48,19 @@ const CreateGroupModal = ({
         return createGroupSchema(d);
     }, []);
 
+    const onClose = () => {
+        if (modalBefore) {
+            setIsModalOpen(false);
+            setterOfModalBefore(true);
+        } else {
+            setIsModalOpen(false);
+        }
+    };
+
     return (
         <Modal
-            data-testid="create-group-modal"
             isModalOpen={isModalOpen}
-            closeModal={() => setIsModalOpen(false)}
+            closeModal={onClose}
             title="Create group"
             submitLabel="Create"
             schema={schema}
@@ -65,5 +75,7 @@ export default CreateGroupModal;
 CreateGroupModal.propTypes = {
     isModalOpen: PropTypes.bool,
     setIsModalOpen: PropTypes.func,
-    reloadData: PropTypes.func
+    reloadData: PropTypes.func,
+    modalBefore: PropTypes.bool,
+    setterOfModalBefore: PropTypes.func
 };